Improve performance of publishing a course with many schools
Before this commit if you select all schools from a provider
with thousand schools this update would take a minute.

The reason is N+1 query and N+1 update statement for each school.

Change to update all does the trick
