+++ app/services/courses/publish_service.rb
@@ -27,10 +27,21 @@ module Courses
     def publish_course
       Course.transaction do
         course.undiscard
-        course.publish_sites
+        publish_sites
         course.publish_enrichment(user)
         course.application_status_open!
       end
+    end
+
+    def publish_sites
+      course.site_statuses
+        .status_new_status
+        .update_all(status: SiteStatus.statuses[:running])
+
+      course.site_statuses
+        .status_running
+        .unpublished_on_ucas
+        .update_all(publish: SiteStatus.publishes[:published])
     end
 
     def send_notification
